backend/introducedog/introducedog/requirements.txt
asgiref==3.2.10
astroid==2.4.2
bcrypt==3.2.0
branca==0.3.1
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
cryptography==3.1
cycler==0.10.0
defusedxml==0.6.0
Django==2.1
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
folium==0.10.1
idna==2.8
isort==5.4.2
Jinja2==2.10.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mysql==0.0.2
mysql-connector-python==8.0.21
mysqlclient==2.0.1
numpy==1.17.3
oauthlib==3.1.0
pandas==0.25.3
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.10.0
pyparsing==2.4.5
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
scipy==1.3.1
seaborn==0.9.0
six==1.13.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.7
wrapt==1.12.1
